ref(naver): refactor oauth2 workflow
