+++ src/Naver/composer.json
@@ -11,9 +11,7 @@
     "require": {
         "php": "^7.2",
         "ext-json": "*",
-        "ext-libxml": "*",
-        "ext-simplexml": "*",
-        "socialiteproviders/manager": "~2.0 || ~3.0"
+        "socialiteproviders/manager": "~3.0"
     },
     "autoload": {
         "psr-4": {
